requirements.txt
pdf2image
pytesseract
openai
# Core Django & Database Packages
django~=5.2
djangorestframework~=3.16.0
drf-spectacular  # OpenAPI schema generation
python-decouple  # For managing environment variables
pytz  # Timezone support
psycopg  # PostgreSQL support
asyncpg  # Async PostgreSQL (if using async DB connections)
pydantic  # Data validation (for API responses)
django-rest-knox~=5.0.2
typing-extensions  # Type hints for Python 3.6
celery~=5.6.2
redis  # Redis support for Celery
django-celery-results  # Celery results backend
gunicorn # WSGI server for production
django-guardian  # Object-level permissions
requests~=2.32.3
django-encrypted-model-fields # Encrypted fields for sensitive data
django-filter~=25.1
python-dateutil~=2.9.0
debugpy # Debugging support for Django
django-cors-headers

django-extensions # Useful extensions for Django development

# AI & ML Frameworks
#groq  # Groq API for AI models (uncomment if needed)
#mistralai  # Mistral AI API (uncomment if needed)
#langchain  # AI-powered reasoning and retrieval
#llama-index  # Indexing & search over AI data
#tantivy  # Fast search engine library
pgvector~=0.4.2  # PostgreSQL vector extension for embeddings


# Web & Deployment Tools
django-environ~=0.12.0
markdown  # Render markdown content
pyyaml  # Parse YAML configs

# Testing & Development
pytest  # Testing framework
pytest-asyncio  # Async test support
pytest-cov  # Test coverage reporting
black  # Code formatting
flake8  # Code linting
mypy  # Static type checking
pytest-django

# Cloud & Storage Services
boto3  # AWS SDK for S3/DigitalOcean Spaces integration
python-magic  # File type detection for uploads

# Security & Virus Scanning
python-clamd  # ClamAV virus scanning client

sentry-sdk~=2.48.0
python-json-logger~=4.0.0
django-prometheus

# GRAPHS
matplotlib
pillow~=12.0.0
scipy~=1.16.3

pyotp~=2.9.0

django_celery_results~=2.6.0
django-qrcode~=0.3
numpy~=2.4.0
django-redis
sendgrid
json-logic~=0.6.3  # JSON Logic expression evaluator for rule engine
tenacity~=9.0.0  # Retry library with exponential backoff
pybreaker~=1.0.1  # Circuit breaker pattern implementation
google-cloud-speech
google-cloud-texttospeech
celery-redbeat  # Redis-based Celery Beat scheduler (kept for potential future use)